On 10853-uplift-load-test-script: 10853 - added scripts changes  AB#10853 AB#10863
