+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.LoadTests/logging/captureLogs.js
@@ -1,0 +1,13 @@
+export function ESSConsoleLog(essResponse){
+    if ( essResponse.status === 200)
+    {
+        let jsonResponse = JSON.parse(essResponse.body);
+        let batchStatusUrl = JSON.stringify(jsonResponse['_links']['exchangeSetBatchStatusUri']['href']);
+    
+        console.log("batchStatusUrl:" + batchStatusUrl + " Status:" + essResponse.status);
+    }
+    else
+    {
+        console.log(" Status:" + essResponse.status); 
+    }
+}
